Fix unable to change race after teamlimit on teamswitch
When you switched team you would not yet be in ..._limit_allowed for the new team when it tried to force you to change race.
This meant you would be forced to change and be stuck without being added to the team limit key and therefore unable to change away.