Added: Level Up Sound & Config
Added the necessary code that will allow players to hear a sound when they level up. For nostalgic purposes the CS:S uses the old level up sound effect.

Also added a configuration option which lets server owners enable / disable this feature in the csgo/cfg/source-python/wcs/config.cfg file. The level up sound effect is disabled by default as it requires the Modified Eventscripts Emulator for WC:S to be installed on the server in order to work for CS:GO.